+++ website/src/shared/context/AuthContext.tsx
@@ -1,22 +1,10 @@
-import React, { createContext, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useConfig } from '@shared/hooks/useConfig';
 import { backendDomain } from '@shared/lib/api';
-
-export interface AuthState {
-  loading: boolean;
-  isAuthenticated: boolean;
-}
-
-export const AuthContext = createContext<AuthState>({
-  loading: true,
-  isAuthenticated: false,
-});
+import { AuthContext, AuthState } from './authContext';
 
 export function AuthProvider({ children }: { children: React.ReactNode }) {
   const { OIDC_ENABLED } = useConfig();
-  // loading starts true only when OIDC is enabled (a fetch is needed).
-  // ConfigProvider returns null while config loads, so OIDC_ENABLED is
-  // already resolved when AuthProvider first renders.
   const [state, setState] = useState<AuthState>(() => ({
     loading: OIDC_ENABLED,
     isAuthenticated: false,
@@ -25,10 +13,15 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   useEffect(() => {
     if (!OIDC_ENABLED) return;
     const controller = new AbortController();
-    fetch(`${backendDomain}/auth/me`, { credentials: 'include', signal: controller.signal })
+    fetch(`${backendDomain}/auth/me`, {
+      credentials: 'include',
+      signal: controller.signal,
+    })
       .then(r => {
-        if (r.status === 200) setState({ loading: false, isAuthenticated: true });
-        else if (r.status === 401) setState({ loading: false, isAuthenticated: false });
+        if (r.status === 200)
+          setState({ loading: false, isAuthenticated: true });
+        else if (r.status === 401)
+          setState({ loading: false, isAuthenticated: false });
         else throw new Error(`Unexpected /auth/me status: ${r.status}`);
       })
       .catch(err => {
