+++ website/src/shared/context/authContext.ts
@@ -1,0 +1,11 @@
+import { createContext } from 'react';
+
+export interface AuthState {
+  loading: boolean;
+  isAuthenticated: boolean;
+}
+
+export const AuthContext = createContext<AuthState>({
+  loading: true,
+  isAuthenticated: false,
+});
