+++ website/src/shared/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { AuthContext } from '@shared/context/AuthContext';
+import { AuthContext } from '@shared/context/authContext';
 
 export function useAuth() {
   return useContext(AuthContext);
